+++ libhanjp/hanjpchar.c
