+++ libhanjp/hanjpchar.h
