+++ libhanjp/libtest.c
