--- libhanjp_new/keyboard/hangul-combination-hj.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<combination id="0">
-    <item first="0x1169" second="0x1161" result="0x116a"/>  <!-- ᅩ   + ᅡ   → ᅪ  -->
-</combination>
